requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.34
python-dotenv==1.0.0
pandas==2.1.1
scikit-learn==1.3.0
